// apps/admin/src/lib/api/generated/models/adminOrderCreate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Hyperlocal Kitchen Platform Admin API
 * OpenAPI spec version: 0.1.0
 */
import type { AdminOrderCreateFulfillmentType } from './adminOrderCreateFulfillmentType';
import type { DeliveryAddress } from './deliveryAddress';
import type { OrderItemCreate } from './orderItemCreate';

export interface AdminOrderCreate {
  pickup_point_id?: string | null;
  delivery_address?: DeliveryAddress | null;
  /**
     * @minItems 1
     * @maxItems 50
     */
  items: OrderItemCreate[];
  fulfillment_type?: AdminOrderCreateFulfillmentType;
  customer_note?: string | null;
  customer_id: string;
}